# reader_concurrency_semaphore: add more layers of defense against OOM