@scylladb-promoter 2022-11-11
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/2969/