@avikivity 2022-11-13
Very good, copying @xemul as someone with experience with OOMs.

Does our query abort cancel I/O?

We'll need unit tests.