@raphaelsc 2022-11-14
we should probably have a metric for OOM kills to differentiate from regular timeouts.