@avikivity 2022-11-14
> > Does our query abort cancel I/O?
> 
> Not directly. After aborting and the exception bubbles up, someone will eventually call `flat_mutation_readr_v2::close()` which will call `close()` on all the file streams and that _is_ supposed to cancel any outstanding I/O AFAIK. @bhalevy, @xemul please ack or correct me.

I think that's good enough. Tough maybe we should create an io_intent when creating the query, and yank it when we abort.