@denesb 2022-11-21
v2 (still RFC):
* expose multiplier param in constructor
* add config option for multiplier and propagate to semaphores
* throw bad alloc instead of using abort source

Change to use abort source for timeout still left in. I can get rid of them or submit separately, not sure if they are improvement or not.