@denesb 2022-11-22
v3 (still RFC):
* Implement the two-layer defense, operating based on soft and hard limits.
* Updated PR title and description.

TODO: testing. Although dev tests are passing, we need more targeted tests for the new functionality.