@scylladb-promoter 2022-11-22
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/3109/