@denesb 2022-11-23
> > > Does our query abort cancel I/O?
> > 
> > 
> > Not directly. After aborting and the exception bubbles up, someone will eventually call `flat_mutation_readr_v2::close()` which will call `close()` on all the file streams and that _is_ supposed to cancel any outstanding I/O AFAIK. @bhalevy, @xemul please ack or correct me.
> 
> I think we need to cancel earlier. It could be (and is even likely) that close() is waiting on a disk read. Waiting for the reads to complete to cancel them won't work!

This ended up not being relevant for this series (as I use bad-alloc now). It is relevant for timeouts though.