@scylladb-promoter 2022-11-29
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/3257/