@denesb 2022-12-09
Perf testing result:
```sh
$ build/release/test/perf/perf_simple_query -c1 -m2G --duration=20
```

BEFORE
```
150891.26 tps ( 62.1 allocs/op,  12.1 tasks/op,   42194 insns/op,        0 errors)
150681.21 tps ( 62.1 allocs/op,  12.1 tasks/op,   42198 insns/op,        0 errors)
151300.27 tps ( 62.1 allocs/op,  12.1 tasks/op,   42192 insns/op,        0 errors)
151514.43 tps ( 62.1 allocs/op,  12.1 tasks/op,   42199 insns/op,        0 errors)
148026.35 tps ( 62.1 allocs/op,  12.1 tasks/op,   42188 insns/op,        0 errors)
149016.53 tps ( 62.1 allocs/op,  12.1 tasks/op,   42186 insns/op,        0 errors)
144832.58 tps ( 62.1 allocs/op,  12.1 tasks/op,   42192 insns/op,        0 errors)
144480.17 tps ( 62.1 allocs/op,  12.1 tasks/op,   42197 insns/op,        0 errors)
143589.74 tps ( 62.1 allocs/op,  12.1 tasks/op,   42183 insns/op,        0 errors)
143058.65 tps ( 62.1 allocs/op,  12.1 tasks/op,   42190 insns/op,        0 errors)

median 149016.53 tps ( 62.1 allocs/op,  12.1 tasks/op,   42186 insns/op,        0 errors)
median absolute deviation: 2497.90
maximum: 151514.43
minimum: 143058.65
```

AFTER
```
150608.84 tps ( 62.1 allocs/op,  12.1 tasks/op,   42796 insns/op,        0 errors)
151653.48 tps ( 62.1 allocs/op,  12.1 tasks/op,   42792 insns/op,        0 errors)
147234.40 tps ( 62.1 allocs/op,  12.1 tasks/op,   42784 insns/op,        0 errors)
142965.79 tps ( 62.1 allocs/op,  12.1 tasks/op,   42781 insns/op,        0 errors)
142516.56 tps ( 62.1 allocs/op,  12.1 tasks/op,   42781 insns/op,        0 errors)
146304.26 tps ( 62.1 allocs/op,  12.1 tasks/op,   42781 insns/op,        0 errors)
141463.78 tps ( 62.1 allocs/op,  12.1 tasks/op,   42783 insns/op,        0 errors)
142699.70 tps ( 62.1 allocs/op,  12.1 tasks/op,   42787 insns/op,        0 errors)
138818.13 tps ( 62.1 allocs/op,  12.1 tasks/op,   42781 insns/op,        0 errors)
142270.98 tps ( 62.1 allocs/op,  12.1 tasks/op,   42791 insns/op,        0 errors)
144169.21 tps ( 62.1 allocs/op,  12.1 tasks/op,   42780 insns/op,        0 errors)
134395.93 tps ( 62.1 allocs/op,  12.1 tasks/op,   42798 insns/op,        0 errors)
140372.47 tps ( 62.1 allocs/op,  12.1 tasks/op,   42786 insns/op,        0 errors)
138218.14 tps ( 62.1 allocs/op,  12.1 tasks/op,   42790 insns/op,        0 errors)
142410.11 tps ( 62.1 allocs/op,  12.1 tasks/op,   42780 insns/op,        0 errors)
146297.01 tps ( 62.1 allocs/op,  12.1 tasks/op,   42782 insns/op,        0 errors)
142184.91 tps ( 62.1 allocs/op,  12.1 tasks/op,   42784 insns/op,        0 errors)
131158.88 tps ( 62.1 allocs/op,  12.1 tasks/op,   42794 insns/op,        0 errors)
137163.27 tps ( 62.1 allocs/op,  12.1 tasks/op,   42793 insns/op,        0 errors)
145068.81 tps ( 62.1 allocs/op,  12.1 tasks/op,   42810 insns/op,        0 errors)

median 142516.56 tps ( 62.1 allocs/op,  12.1 tasks/op,   42781 insns/op,        0 errors)
median absolute deviation: 3698.43
maximum: 151653.48
minimum: 131158.88
```

There is a 1.4% increase in instruction count / op. I will v5 which includes a small change which helped me shave it down to this value (was slightly higher).