@avikivity 2022-12-11
500 extra instructions is a lot. Where does it come from?