@avikivity 2022-12-11
Code looks good, but where did those 500 instructions come from? Maybe bisect will tell.