@denesb 2022-12-12
> Code looks good, but where did those 500 instructions come from? Maybe bisect will tell.

I don't know, so I'll have to bisect.

Also, in release mode the multishard fuzzy test times out, possibly a deadlock was introduced.

So a new version will be coming.