@avikivity 2022-12-20
> Update: I started working on a manual test to confirm the mechanism does its job on a real read workload. I wanted to observe `master` OOMing or at least consuming a lot of memory visible in metrics and this PR capping this consumption at the desired level. Instead of this so far I got some assert failures that I need to look into. I also need to consider why my existing unit test didn't catch this. It would be good to have an automated regression test reproducing these assert failures but this is traditionally hard to do because of the amount of data required to reproduce read OOMs. I will maybe have success making such a test have an acceptable duration with `scylla sstable write`.

We can artificially limit memory size for that.